requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.23.6
websockets==12.0
faiss-cpu==1.7.4
numpy==1.26.4
python-dotenv==1.0.1
